Full commit
This commit consist of three different commits which I had to put as a single one because of versioning issues.

feature: basic GUI with screen switching and simple JTable prototype. Basic arhitecture of the application

features: Implemented a JPanel that allows the user adjust settings, made a json file which has the ability of loading/saving user settings, Improved the hadoop files loading time by creating an index of files, Refactored a bit of code, Added a JDialogBox for showing the progress of a file/directory copy

features: fixed the settings issue, fixed a bug when loading folders with only one file, fixed a bug when loading a folder multiple time, fixed the settings issues so they are applied immediately, added the settings as a jmenubar item